pyproject.toml
[tool.poetry]
name = "kursovaya-django"
version = "0.1.0"
description = ""
authors = ["Annabse27 <annabse27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
redis = "^5.0.8"
django-celery-beat = "^2.7.0"
psycopg2-binary = "^2.9.9"
django-apscheduler = "^0.6.2"
python-dotenv = "^1.0.1"
django-celery-results = "^2.5.1"
django-allauth = "^64.2.0"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
cryptography = "^43.0.0"
pillow = "^10.4.0"
django-redis = "^5.4.0"
slugify = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
